refactor: casting struct variables

+++ engine/texture.cpp
@@ -29,7 +29,7 @@ struct texture texture_load(const char* path)
 
 	free(image);
 
-	struct texture t = { texture_handle, width, height };
+	struct texture t = { texture_handle, (unsigned short)width, (unsigned short)height };
 	return t;
 }
